ai/requirements.txt
dash==2.17.1
dash-bootstrap-components==1.5.0
plotly==5.17.0
python-dotenv==1.0.0
gunicorn==21.2.0
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.2.39
openai==1.10.0
langchain-community==0.0.13
pydantic==2.5.3

# Document processing dependencies
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
unstructured==0.12.4
markdown==3.5.2
beautifulsoup4==4.12.2
lxml==4.9.4
pymupdf==1.23.26